+++ src/main/java/com/vstu/department/model/EmployeeParameter.java
@@ -1,12 +1,18 @@
 package com.vstu.department.model;
 
+import javax.persistence.AttributeOverride;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+import javax.validation.constraints.NotNull;
+
 import com.vstu.department.model.enums.AnketaParameterStatusType;
+
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
-
-import javax.persistence.*;
-import javax.validation.constraints.NotNull;
 
 @Getter
 @Setter
@@ -30,7 +36,7 @@ public class EmployeeParameter extends PersistentEntity {
 
     @NotNull
     @Column(name = "emp_count")
-    private Long count;
+    private Double count;
 
     @NotNull
     @Column(name = "emp_coefficient")
